requirements.txt
annotated-types==0.7.0
asgiref==3.7.2
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
blinker==1.7.0
Brlapi==0.8.5
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.6
cloud-init==24.4.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
dotenv==0.9.9
duplicity==2.1.4
fasteners==0.18
filelock==3.18.0
Flask==3.0.2
fsspec==2025.3.2
generativeai==0.0.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
hf-xet==1.1.0
httplib2==0.22.0
huggingface-hub==0.31.1
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
Mako==1.3.2.dev0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
netaddr==0.8.0
oauthlib==3.2.2
olefile==0.46
packaging==25.0
paramiko==2.12.0
pexpect==4.9.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyasyncore==1.0.2
pycairo==1.25.1
pycparser==2.22
pycups==2.0.1
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.17.2
PyGObject==3.48.2
pyinotify==0.9.6
PyJWT==2.7.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.2.3
pyperclip==1.9.0
pyrsistent==0.20.0
pyserial==3.5
pytesseract==0.3.13
python-apt==2.7.7+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.1.0
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9.1
setuptools==68.1.2
simplejson==3.19.2
six==1.16.0
systemd-python==235
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.4.0
usb-creator==0.3.16
wadllib==1.3.6
watchdog==6.0.0
Werkzeug==3.0.1
wheel==0.42.0
xdg==5
xkit==0.0.0
